Move more types to debugger from shared
