+++ packages/debugger/src/inspector/index.ts
@@ -1,11 +1,10 @@
 import { Accessor, createEffect, onCleanup, untrack } from 'solid-js'
 import { throttle, scheduleIdle } from '@solid-primitives/scheduled'
-import { NodeID, EncodedValue, ValueNodeId } from '@solid-devtools/shared/graph'
 import { warn } from '@solid-devtools/shared/utils'
-import { DebuggerEventHub } from '../plugin'
-import { walkSolidRoot } from '../roots'
-import { Core, Solid } from '../types'
-import { makeSolidUpdateListener } from '../update'
+import { DebuggerEventHub } from '../main/plugin'
+import { walkSolidRoot } from '../main/roots'
+import { Core, EncodedValue, NodeID, Solid, ValueNodeId } from '../main/types'
+import { makeSolidUpdateListener } from '../main/update'
 import { NodeIDMap, encodeValue } from './serialize'
 import { observeStoreNode, StoreUpdateData } from './store'
 import { clearOwnerObservers, collectOwnerDetails, ValueNode, ValueNodeMap } from './inspector'
