+++ packages/debugger/src/locator/findComponent.ts
@@ -1,6 +1,6 @@
-import { Mapped, NodeID } from '@solid-devtools/shared/graph'
 import { LOCATION_ATTRIBUTE_NAME } from '@solid-devtools/shared/variables'
 import { isWindows } from '@solid-primitives/platform'
+import { Mapped, NodeID } from '../types'
 import { ElementLocation } from './goToSource'
 
 export type LocatorComponent = {
