+++ packages/debugger/src/locator/index.ts
@@ -4,7 +4,6 @@ import { createKeyHold, KbdKey } from '@solid-primitives/keyboard'
 import { onRootCleanup } from '@solid-primitives/utils'
 import { createSimpleEmitter } from '@solid-primitives/event-bus'
 import { atom, defer, makeHoverElementListener } from '@solid-devtools/shared/primitives'
-import { Mapped, NodeID } from '@solid-devtools/shared/graph'
 import { findLocatorComponent, getLocationFromElement, LocatorComponent } from './findComponent'
 import {
   getFullSourceCodeData,
@@ -13,9 +12,10 @@ import {
   TargetIDE,
   TargetURLFunction,
 } from './goToSource'
-import { createInternalRoot } from '../utils'
-import { enableRootsAutoattach } from '../roots'
+import { createInternalRoot } from '../main/utils'
+import { enableRootsAutoattach } from '../main/roots'
 import { attachElementOverlay } from './ElementOverlay'
+import { Mapped, NodeID } from '../types'
 
 export type { TargetIDE, TargetURLFunction } from './goToSource'
 export type { LocatorComponent } from './findComponent'
@@ -26,6 +26,11 @@ export type LocatorOptions = {
   /** Holding which key should enable the locator overlay? */
   key?: KbdKey
 }
+
+export type HighlightElementPayload =
+  | { rootId: NodeID; nodeId: NodeID }
+  | { elementId: string }
+  | null
 
 export type ClickMiddleware = (
   e: MouseEvent,
@@ -105,9 +110,7 @@ export function createLocator({
     return comp?.id
   })
 
-  function setPluginHighlightTarget(
-    data: { elementId: NodeID } | { rootId: NodeID; nodeId: NodeID } | null | undefined,
-  ) {
+  function setPluginHighlightTarget(data: HighlightElementPayload) {
     if (!data) return pluginTarget(null)
     // highlight component
     if ('nodeId' in data) {
