+++ packages/debugger/src/main/roots.ts
@@ -1,6 +1,5 @@
 import { createEffect, onCleanup, untrack } from 'solid-js'
 import { throttle } from '@solid-primitives/scheduled'
-import { NodeID, NodeType } from '@solid-devtools/shared/graph'
 import { untrackedCallback } from '@solid-devtools/shared/primitives'
 import { warn } from '@solid-devtools/shared/utils'
 import { ComputationUpdateHandler, WalkerResult, walkSolidTree } from './walker'
@@ -18,7 +17,8 @@ import {
   INTERNAL,
 } from './utils'
 import { makeCreateRootListener } from './update'
-import { Core, DebuggerContext, Solid } from './types'
+import { Core, DebuggerContext, NodeID, Solid } from './types'
+import { NodeType } from './constants'
 
 const RootMap: Record<NodeID, (inspectedId?: NodeID) => WalkerResult | null> = {}
 export const walkSolidRoot = (rootId: NodeID, inspectedId?: NodeID): WalkerResult | null => {
