+++ packages/debugger/src/main/test/utils.test.ts
@@ -1,5 +1,4 @@
 import { describe, it, expect } from 'vitest'
-import { NodeType } from '@solid-devtools/shared/graph'
 import {
   createComponent,
   createComputed,
@@ -9,8 +8,9 @@ import {
   createRoot,
   createSignal,
 } from 'solid-js'
-import { getFunctionSources, getOwner, getOwnerType, onDispose } from '../src/utils'
-import { Solid } from '../src/types'
+import { getFunctionSources, getOwner, getOwnerType, onDispose } from '../utils'
+import { Solid } from '../types'
+import { NodeType } from '../constants'
 
 describe('getOwnerType', () => {
   it('identifies Component', () => {
