+++ packages/debugger/src/main/utils.ts
@@ -7,10 +7,10 @@ import {
 } from 'solid-js'
 import { Emit } from '@solid-primitives/event-bus'
 import { throttle } from '@solid-primitives/scheduled'
-import { NodeType, NodeID } from '@solid-devtools/shared/graph'
 import { trimString } from '@solid-devtools/shared/utils'
 import { DEV as STORE_DEV } from 'solid-js/store'
-import { Core, DebuggerContext, Solid } from './types'
+import { Core, DebuggerContext, NodeID, Solid, ValueNodeId, ValueNodeType } from './types'
+import { NodeType } from './constants'
 
 export const getOwner = _getOwner as () => Solid.Owner | null
 
@@ -86,6 +86,14 @@ export const getOwnerType = (o: Readonly<Solid.Owner>): NodeType => {
 
 let LAST_ID = 0
 export const getNewSdtId = (): NodeID => (LAST_ID++).toString(36)
+
+export const getValueNodeId = <T extends ValueNodeType>(
+  type: T,
+  id: T extends 'value' ? undefined : NodeID | string,
+): ValueNodeId => {
+  if (type === 'value') return 'value'
+  return `${type}:${id}` as ValueNodeId
+}
 
 export function markOwnerName(o: Solid.Owner): string {
   if (o.sdtName !== undefined) return o.sdtName
