+++ packages/debugger/tsup.config.ts
@@ -4,4 +4,5 @@ export default defineConfig({
   extension: 'ts',
   server: true,
   jsx: true,
+  additionalEntries: ['types'],
 })
