+++ packages/extension/package.json
@@ -11,7 +11,7 @@
     "build": "npm run clean && vite build && node scripts/zip.cjs",
     "clean": "rimraf dist dist.zip",
     "test": "vitest",
-    "typecheck": "tsc --noEmit"
+    "typecheck": "tsc --noEmit --paths null"
   },
   "devDependencies": {
     "@crxjs/vite-plugin": "^1.0.14",
