+++ packages/frontend/tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "paths": {
-      "@solid-devtools/shared/*": ["../shared/src/*"],
       "@/*": ["./src/*"]
     }
   },
