+++ packages/logger/package.json
@@ -48,7 +48,7 @@
     "dev": "tsup --watch",
     "build": "tsup",
     "test": "echo \"NOOP test\"",
-    "typecheck": "tsc --noEmit"
+    "typecheck": "tsc --noEmit --paths null"
   },
   "devDependencies": {
     "solid-js": "^1.6.0",
