+++ packages/shared/src/index.ts
@@ -1,4 +1,3 @@
-export * as graph from './graph'
 export * as primitives from './primitives'
 export * as utils from './utils'
 export * as variables from './variables'
