+++ packages/transform/package.json
@@ -42,7 +42,7 @@
     "dev": "tsup --watch",
     "build": "tsup",
     "test": "vitest",
-    "typecheck": "tsc --noEmit"
+    "typecheck": "tsc --noEmit --paths null"
   },
   "devDependencies": {
     "@babel/core": "^7.18.13",
